requirements.txt
streamlit
google-genai
python-dotenv
mcp
fastapi
uvicorn
# sseclient-py